Added Visitor support in File class and subclasses
The generic method accept was added in the File class as abstract and as
each subclass as an implementation.

By parameterizing the return type and the generic visitor argument each
accept method returns the corresponding type to the Visitor object
parameter passed as argument.

Each method will call the corresponding Visitor visit() with a reference
to itself as argument to allow class diferentation as expected in the
Visitor pattern without specifying the return type