Changed DirectoryVisitor to return Directories.
DirectoryVisitor previously returned Boolean's, which would only enable us to
verify if the given argument is a directory or not. If we wanted to actually get
the directory we would need to use a cast.

The aim of the change is to avoid casts, which is always a good thing, while
keeping the same information as we would using Boolean's.