Added remove() getSize() and both listings
The remove() method is basically deleting the file kept by the directory and the
references kept by it. getSize() is pretty intuitive.

It was added both simple and complex listings. There was also created a generic
listing method that both simple and complex listings methods use to keep the
code cleaner. For further explanation see the JavaDoc of the method, which has a
pretty fair explanation.