Added RootDirectory class fixed #22
The problem was that the fenixframework adds objects to both sides of the
relation, inducing undesirable behaviour -- when setting the rootDirectory's
parent to itself, the directory would be added as a member of the files
contained by itself.

The way the issue was fixed was by creating a RootDirectory class to override
directory behaviour. The ideal way would be to create an inner class of the
FileSystem, but the everlasting awesomeness of the fenix framework doesn't
allows us to do it. Therefore the .dml had to be altered and so on...

There were also added some new lines to the main class to illustrate the new
behaviour.