Added builders closes #20
In deep hierarchical class systems, builders can be helpful dealing with the
exponenciality of constructors, namely, when there are, in top level classes, a
lot of constructors -- due to, maybe, a lot of optional arguments -- improving
extensibility.
They are also useful to separate the building logic of an object from the
class itself, althought in our case, it doesn't really apply.

The syntax for the builders is a little cryptic, although its usage is pretty
straight forward.
For example, if we were to create Directory we would do:

Directory d = Directory.DirectoryBuilder.create()
                                        .setX(...)
                                        .setY(...)
                                        .setZ(...)
                                        .build()

NOTE: The first line of the building process:
Directory.DirectoryBuilder.create() can be eventually squashed into a single
function call using static imports.